requirements.txt
torch>=2.0
torchvision>=0.15
scikit-learn>=1.3
Pillow>=10.0
tqdm>=4.66
fastapi>=0.110
uvicorn>=0.29
python-multipart>=0.0.9
